+++ python/brain/__init__.py
@@ -13,7 +13,7 @@ import netrc
 import warnings
 import yaml
 from brain.core.exceptions import BrainError, BrainUserWarning
-from brain.utils.general.general import merge
+from brain.utils.general.general import merge, get_yaml_loader
 
 # Set the Brain version
 __version__ = '0.1.2dev'
@@ -94,10 +94,10 @@ class BrainConfig(object):
     def _load_defaults(self):
         ''' Load the Brain config yaml file '''
 
-        config = yaml.load(open(os.path.join(os.path.dirname(__file__), 'data/brain.yml')))
+        config = yaml.load(open(os.path.join(os.path.dirname(__file__), 'data/brain.yml')), Loader=get_yaml_loader())
         user_config_path = os.path.expanduser('~/.brain/brain.yml')
         if os.path.exists(user_config_path):
-            config = merge(yaml.load(open(user_config_path)), config)
+            config = merge(yaml.load(open(user_config_path), Loader=get_yaml_loader()), config)
 
         # update any matching Config values
         for key, value in config.items():